+++ source/Switchery/index.js
@@ -1,0 +1,6 @@
+/**
+ * @since 2016-09-06 09:13
+ * @author vivaxy
+ */
+
+export default from './Switchery';